Move UserDefaults mock to separate file
